+++ examples/tracing-http/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "traceresponse"
+name = "tracing-http"
 version = "0.1.0"
 edition = "2021"
 license = "Apache-2.0"
